backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "todo-web-backend"
version = "1.0.0"
description = "FastAPI backend for Todo Web Application with Better Auth"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "Todo App Developer"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: FastAPI",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    # FastAPI and Server
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    # Database
    "sqlmodel>=0.0.22",
    "psycopg2-binary>=2.9.9",
    "asyncpg>=0.30.0",
    # Authentication
    "httpx>=0.27.0",
    # AI and MCP
    "openai>=1.50.0",
    "python-mcp>=0.1.0",
    # Utilities
    "python-dotenv>=1.0.1",
    "pydantic[email]>=2.9.0",
    "pydantic-settings>=2.6.0",
    # CORS
    "python-multipart>=0.0.12",
    # Date handling
    "python-dateutil>=2.9.0",
    "pyjwt>=2.10.1",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
asyncio_mode = "auto"
